Extract ingest-cursor reconciliation into a mixin (WS5 Seam 4)
Move the ingest-cursor reconciliation machinery and the stable replay-identity
primitives it relies on out of LCMEngine into a ReconcileMixin in the new
reconcile.py, and mix it in (`class LCMEngine(ContextEngine, ReconcileMixin)`).

Moved (20 methods): the tool-call/message replay-identity primitives, the
persisted-output durable-replay evidence helpers, and the reconciliation
orchestration (_find_reconciled_cursor_for_store_tail,
_reconcile_ingest_cursor_from_store, _get_store_id_map_for_messages, etc.). They
keep running bound to the engine instance, so _store, _session_id, _config and
the safety-critical _ingest_cursor stay on the engine and resolve via self;
_ingest_cursor is still written by _ingest_messages from the reconciler's return.

Method bodies are byte-identical to their pre-move form except two mechanical,
behaviour-preserving adjustments:
- The two tool-call-identity staticmethods call the identity canonicalizer via
  the mixin class (ReconcileMixin._canonicalize_tool_call_identity_value) instead
  of LCMEngine, to avoid an engine<->reconcile import cycle. staticmethod
  resolution is identical (LCMEngine inherits it from the mixin).
- _PRESERVED_OBJECTIVE_CONTEXT_PREFIX moves to reconcile.py (used by the
  reconciliation scan) and is re-imported into engine.py for the other users, so
  no code that reads it changes.

engine.py shrinks by ~1105 lines and drops six now-exclusive ingest_protection/
externalize imports (moved to reconcile.py). No call sites and no tests change.

Behaviour-preserving refactor (WS5 engine decomposition). Rollback: git revert
this commit re-inlines the methods; no data or state migration involved.

Bases order: the mixin precedes ContextEngine so the LCM decomposition mixins
resolve consistently across seams (required for CompactionMixin, which overrides
the compress/should_compress protocol methods; harmless for private-only mixins).
